+++ pytest.ini
@@ -23,6 +23,7 @@ markers =
     gpu: mark a test that requires GPU.
     integration: mark an integration test
     onnx_coverage: ONNX coverage test
+    garbage_expected: this teast leaks ndarray references. The tested functionality is broken.
 
 env =
     MXNET_HOME=tests/data
