+++ tests/python/unittest/test_gluon_data.py
@@ -225,6 +225,7 @@ def test_image_list_dataset_handle(prepare_record):
         assert label == 0
 
 @with_seed()
+@pytest.mark.garbage_expected
 def test_list_dataset():
     for num_worker in range(0, 3):
         data = mx.gluon.data.DataLoader([([1,2], 0), ([3, 4], 1)], batch_size=1, num_workers=num_worker)
